requirements.txt
flask
gunicorn
pycryptodome
werkzeug
pandas
fastapi
uvicorn
pillow
numpy
python-multipart
sqlalchemy